Relocate Profile Type
