+++ src/features/AppCoordinator.ts
@@ -19,7 +19,6 @@ import { IPatientService } from '@covid/core/patient/PatientService';
 import { IContentService } from '@covid/core/content/ContentService';
 import NavigatorService from '@covid/NavigatorService';
 import Analytics, { events } from '@covid/core/Analytics';
-import { Profile } from '@covid/components/Collections/ProfileList';
 import { PatientData } from '@covid/core/patient/PatientData';
 import editProfileCoordinator from '@covid/features/multi-profile/edit-profile/EditProfileCoordinator';
 import store from '@covid/core/state/store';
@@ -36,6 +35,7 @@ import { UserResponse } from '@covid/core/user/dto/UserAPIContracts';
 import { Coordinator, IEditableProfile, ISelectProfile } from '@covid/core/Coordinator';
 import dietStudyPlaybackCoordinator from '@covid/features/diet-study-playback/DietStudyPlaybackCoordinator';
 import { IDietScoreRemoteClient } from '@covid/core/diet-score/DietScoreApiClient';
+import { Profile } from '@covid/core/profile/ProfileService';
 
 type ScreenName = keyof ScreenParamList;
 type ScreenFlow = {
